LayoutTests/fast/text-extraction/debug-text-extraction-interactions-expected.txt

PASS clickSucceeded is true
PASS clickCount.textContent is "1"
PASS textInputSucceeded is true
PASS textField.value is "bar"
PASS selectMenuItemSucceeded is true
PASS select.value is "Three"
PASS selectTextSucceeded is true
PASS getSelection().toString() is "Subject"
PASS invalidActionSucceeded is false
PASS successfullyParsed is true

TEST COMPLETE
Test
Subject

Hello world.

1
